web/requirements.txt
python-dotenv
boto3
openpyxl
requests
pandas
selenium